Revert "List Controls Styling"
