--- src/components/AddItem.css
@@ -1,151 +1,0 @@
-/** */
-
-.addItem__form__hasItems {
-	display: flex;
-	flex-direction: column;
-}
-
-.addItem__section__hasItems {
-	display: flex;
-	justify-content: space-between;
-	align-content: center;
-}
-
-.addItem__label__hasItems {
-	display: flex;
-	align-content: center;
-	gap: 12px;
-}
-
-.addItem__modifiers__hasItems {
-	display: flex;
-	flex-wrap: wrap;
-	width: 100%;
-	gap: 20px;
-	margin: 24px 0px 24px 0px;
-	justify-content: start;
-}
-
-.addItem__select__hasItems {
-	margin: auto;
-	border-radius: 10px;
-}
-
-#userList__hasItems {
-	width: 180px;
-}
-
-.label__select__hasItems {
-	display: flex;
-	flex-wrap: wrap;
-}
-
-#addInput__hasItems {
-	width: 100%;
-}
-
-.input__hasItems {
-	width: 100%;
-}
-
-.p__hasItems {
-	width: auto;
-	padding-top: 2px;
-}
-
-.btn__submit__hasItems {
-	margin: auto;
-	margin-left: 0.5rem;
-	background: var(--color-accent);
-	font-size: 1.6rem;
-	/* width: 150px; */
-}
-#submit {
-	margin: 1rem auto;
-	padding: 1px;
-	height: 30px;
-	width: 150px;
-	border-radius: 2rem;
-	background: var(--color-accent);
-}
-
-.addItem__form {
-	display: flex;
-	flex-direction: column;
-}
-
-.addItem__label {
-	margin: 1rem;
-}
-
-.addItem__label > * {
-	margin: 0.5rem;
-}
-
-.addItem__input {
-	font-size: 16px;
-	padding: 0.5em;
-	color: var(--color-text);
-	text-align: center;
-}
-
-.addItem__select {
-	margin-left: 1rem;
-	border-radius: 10px;
-}
-
-.addItem__section {
-	display: flex;
-	flex-direction: column;
-	align-items: center;
-}
-
-/* long list names? */
-.addItem__select_listNames {
-	width: fit-content;
-	padding-right: 3em;
-}
-
-.addItem__select_frequency {
-	width: 180px;
-}
-
-@media screen and (max-width: 859px) {
-	.addItem__modifiers__hasItems {
-		justify-content: center;
-	}
-}
-
-@media screen and (max-width: 768px) {
-	#addInput__hasItems {
-		display: flex;
-		flex-direction: column;
-		margin-left: 10%;
-		margin-right: 10%;
-	}
-	.addItem__section__hasItems {
-		padding-bottom: 12px;
-		justify-content: center;
-		margin: auto;
-		width: 100%;
-	}
-	.addItem__modifiers__hasItems {
-		flex-direction: column;
-		justify-content: center;
-		align-content: center;
-		margin: auto;
-	}
-	.p__hasItems {
-		text-align: center;
-	}
-	.addItem__label__hasItems {
-		display: flex;
-		flex-direction: column;
-	}
-	/* }
-@media screen and (max-width: 768px) { */
-
-	.addItem__label {
-		margin: 0 0 1em 0;
-	}
-}
