--- src/components/Filter.css
@@ -1,139 +1,0 @@
-.Filter {
-	outline: var(--color-accent) solid 2px;
-	padding: 20px;
-	border-radius: 10px;
-}
-
-.Filter.wide {
-	padding: 5px 20px 20px 20px;
-	margin-bottom: 3rem;
-}
-
-fieldset,
-.urgency__label,
-details {
-	margin-top: 24px;
-	margin-bottom: 24px;
-}
-
-.Filter__heading {
-	font-size: 2rem;
-	font-weight: 600;
-	color: var(--color-input);
-}
-details.Filter__heading {
-	cursor: pointer;
-}
-
-.search__container {
-	display: flex;
-	width: 100%;
-	gap: 4px;
-	margin-top: 12px;
-	margin-bottom: 12px;
-}
-
-.search__input {
-	margin: auto;
-	height: 45px;
-	width: 100%;
-	font-size: 1.6rem;
-}
-
-fieldset {
-	display: flex;
-	justify-content: space-between;
-	padding: 24px;
-	border: var(--color-accent) dashed 1px;
-	border-radius: 10px;
-}
-
-.customSearch__label {
-	display: flex;
-	justify-content: center;
-	align-content: center;
-	gap: 6px;
-	width: 100%;
-}
-
-.customSearch__label > input {
-	width: 100%;
-	height: 30px;
-	margin: auto 14px auto 14px;
-	font-size: 1.6rem;
-}
-
-.btn__clearAll {
-	background: var(--color-accent);
-	font-size: 1.6rem;
-}
-
-.urgency__label {
-	display: flex;
-	justify-content: start;
-	gap: 12px;
-}
-
-.urgency__label > select {
-	border-radius: 10px;
-}
-
-@media screen and (max-width: 768px) {
-	fieldset,
-	.urgency__label {
-		flex-direction: column;
-	}
-
-	.urgency__label {
-		justify-content: center;
-		align-content: center;
-		text-align: center;
-	}
-
-	.urgency__label > select {
-		margin: auto;
-	}
-
-	.Filter__heading {
-		text-align: center;
-	}
-
-	/* mobile collapse */
-	summary {
-		cursor: pointer;
-		display: flex;
-		justify-content: space-evenly;
-		align-items: baseline;
-	}
-
-	.container__btn__clearAll {
-		display: flex;
-		justify-content: center;
-	}
-}
-
-@media screen and (max-width: 640px) {
-	.search__container {
-		flex-direction: column;
-		justify-content: center;
-		align-items: center;
-		text-align: center;
-	}
-	.search__input {
-		text-align: center;
-	}
-}
-
-@media screen and (max-width: 400px) {
-	/* mobile collapse */
-	summary {
-		cursor: pointer;
-		display: flex;
-		flex-direction: column;
-		justify-content: space-between;
-		align-items: center;
-	}
-	summary > .container__btn__clearAll {
-		margin-top: 1.6rem;
-	}
-}
