--- src/components/ListTitle.css
@@ -1,4 +1,0 @@
-form {
-	display: flex;
-	justify-content: center;
-}
